perf(cms): refine dashboard navigation loading
