+++ apps/cms/src/components/nav/nav-settings.tsx
@@ -22,11 +22,7 @@ import {
 import { cn } from "@marble/ui/lib/utils";
 import { useQueryClient } from "@tanstack/react-query";
 import Link from "next/link";
-import {
-  useParams,
-  useRouter,
-  useSelectedLayoutSegments,
-} from "next/navigation";
+import { useParams, useSelectedLayoutSegments } from "next/navigation";
 import { prefetchDashboardRoute } from "@/lib/dashboard-prefetch";
 import { useWorkspace } from "@/providers/workspace";
 import { workspacePath } from "@/utils/workspace/url";
@@ -87,7 +83,6 @@ const developerItems = [
 
 export function NavSettings() {
   const segments = useSelectedLayoutSegments();
-  const router = useRouter();
   const params = useParams<{ workspace: string }>();
   const queryClient = useQueryClient();
   const { activeWorkspace } = useWorkspace();
@@ -98,9 +93,6 @@ export function NavSettings() {
     .join("/");
   const isActive = (url: string) => activeUrl === url;
   const prefetchRoute = (url: string) => {
-    const href = workspacePath(params.workspace, url);
-    router.prefetch(href);
-
     if (activeWorkspace?.id) {
       prefetchDashboardRoute(queryClient, activeWorkspace.id, url).catch(
         () => undefined
