+++ apps/cms/src/components/shared/page-loader.tsx
@@ -1,12 +1,10 @@
-"use client";
-
-import { SpinnerIcon } from "@phosphor-icons/react";
+import { ActivityIndicator } from "@/components/ui/activity-indicator";
 
 function PageLoader() {
   return (
     <div aria-busy="true" className="grid h-full w-full place-content-center">
       <div className="p-2">
-        <SpinnerIcon className="size-5 animate-spin transition" />
+        <ActivityIndicator />
       </div>
     </div>
   );
