+++ apps/cms/src/lib/queries/dashboard/usage.ts
@@ -1,0 +1,214 @@
+import "server-only";
+
+import { db } from "@marble/db";
+import { UsageEventType } from "@marble/db/browser";
+import { addDays, format, startOfDay, subDays, subHours } from "date-fns";
+import type { UsageDashboardData } from "@/types/dashboard";
+
+const CHART_DAYS = 30;
+
+export async function getDashboardUsageMetrics(
+  workspaceId: string
+): Promise<UsageDashboardData> {
+  const now = new Date();
+  const today = startOfDay(now);
+  const chartStart = subDays(today, CHART_DAYS - 1);
+  const previousPeriodStart = subDays(chartStart, CHART_DAYS);
+
+  const [apiEvents, apiPrevPeriodCount, apiTotalCount] = await Promise.all([
+    db.usageEvent.findMany({
+      where: {
+        workspaceId,
+        type: UsageEventType.api_request,
+        createdAt: { gte: chartStart },
+      },
+      select: { createdAt: true },
+    }),
+    db.usageEvent.count({
+      where: {
+        workspaceId,
+        type: UsageEventType.api_request,
+        createdAt: {
+          gte: previousPeriodStart,
+          lt: chartStart,
+        },
+      },
+    }),
+    db.usageEvent.count({
+      where: {
+        workspaceId,
+        type: UsageEventType.api_request,
+      },
+    }),
+  ]);
+
+  const chartBuckets = new Map<string, number>();
+  for (let i = 0; i < CHART_DAYS; i += 1) {
+    const date = addDays(chartStart, i);
+    chartBuckets.set(format(date, "yyyy-MM-dd"), 0);
+  }
+  for (const event of apiEvents) {
+    const key = format(startOfDay(event.createdAt), "yyyy-MM-dd");
+    chartBuckets.set(key, (chartBuckets.get(key) ?? 0) + 1);
+  }
+
+  const apiChart = Array.from(chartBuckets.entries()).map(
+    ([dateKey, count]) => ({
+      date: dateKey,
+      label: format(new Date(dateKey), "MMM d"),
+      value: count,
+    })
+  );
+
+  const apiLastPeriodCount = apiChart.reduce(
+    (acc, curr) => acc + curr.value,
+    0
+  );
+  const apiChange =
+    apiPrevPeriodCount === 0
+      ? apiLastPeriodCount > 0
+        ? 100
+        : 0
+      : ((apiLastPeriodCount - apiPrevPeriodCount) / apiPrevPeriodCount) * 100;
+
+  const webhookChartStart = subDays(today, CHART_DAYS - 1);
+  const [
+    webhookTotal,
+    webhookWeek,
+    webhookDay,
+    webhookTopEndpoint,
+    webhookEvents,
+    mediaTotals,
+    mediaLast30,
+    mediaLastUpload,
+    recentMediaUploads,
+  ] = await Promise.all([
+    db.usageEvent.count({
+      where: { workspaceId, type: UsageEventType.webhook_delivery },
+    }),
+    db.usageEvent.count({
+      where: {
+        workspaceId,
+        type: UsageEventType.webhook_delivery,
+        createdAt: { gte: subDays(now, 6) },
+      },
+    }),
+    db.usageEvent.count({
+      where: {
+        workspaceId,
+        type: UsageEventType.webhook_delivery,
+        createdAt: { gte: subHours(now, 24) },
+      },
+    }),
+    db.usageEvent.groupBy({
+      by: ["endpoint"],
+      where: {
+        workspaceId,
+        type: UsageEventType.webhook_delivery,
+        endpoint: { not: null },
+      },
+      _count: { endpoint: true },
+      orderBy: { _count: { endpoint: "desc" } },
+      take: 1,
+    }),
+    db.usageEvent.findMany({
+      where: {
+        workspaceId,
+        type: UsageEventType.webhook_delivery,
+        createdAt: { gte: webhookChartStart },
+      },
+      select: { createdAt: true },
+    }),
+    db.usageEvent.count({
+      where: { workspaceId, type: UsageEventType.media_upload },
+    }),
+    db.usageEvent.count({
+      where: {
+        workspaceId,
+        type: UsageEventType.media_upload,
+        createdAt: { gte: subDays(now, 29) },
+      },
+    }),
+    db.usageEvent.findFirst({
+      where: { workspaceId, type: UsageEventType.media_upload },
+      orderBy: { createdAt: "desc" },
+      select: { createdAt: true },
+    }),
+    db.media.findMany({
+      where: { workspaceId },
+      orderBy: { createdAt: "desc" },
+      take: 10,
+      select: {
+        id: true,
+        name: true,
+        size: true,
+        alt: true,
+        createdAt: true,
+        type: true,
+        url: true,
+        mimeType: true,
+        width: true,
+        height: true,
+        duration: true,
+        blurHash: true,
+      },
+    }),
+  ]);
+
+  const webhookChartBuckets = new Map<string, number>();
+  for (let i = 0; i < CHART_DAYS; i += 1) {
+    const date = addDays(webhookChartStart, i);
+    webhookChartBuckets.set(format(date, "yyyy-MM-dd"), 0);
+  }
+  for (const event of webhookEvents) {
+    const key = format(startOfDay(event.createdAt), "yyyy-MM-dd");
+    webhookChartBuckets.set(key, (webhookChartBuckets.get(key) ?? 0) + 1);
+  }
+
+  const webhookChart = Array.from(webhookChartBuckets.entries()).map(
+    ([dateKey, count]) => ({
+      date: dateKey,
+      label: format(new Date(dateKey), "MMM d"),
+      value: count,
+    })
+  );
+
+  return {
+    api: {
+      totals: {
+        total: apiTotalCount,
+        lastPeriod: apiLastPeriodCount,
+        changePercentage: Math.round(apiChange * 100) / 100,
+      },
+      chart: apiChart,
+    },
+    webhooks: {
+      total: webhookTotal,
+      last7Days: webhookWeek,
+      last24Hours: webhookDay,
+      topEndpoint: webhookTopEndpoint[0]?.endpoint ?? null,
+      topEndpointCount: webhookTopEndpoint[0]?._count.endpoint ?? 0,
+      chart: webhookChart,
+    },
+    media: {
+      total: mediaTotals,
+      last30Days: mediaLast30,
+      totalSize: recentMediaUploads.reduce((sum, media) => sum + media.size, 0),
+      lastUploadAt: mediaLastUpload?.createdAt.toISOString() ?? null,
+      recentUploads: recentMediaUploads.map((media) => ({
+        id: media.id,
+        name: media.name,
+        size: media.size,
+        alt: media.alt,
+        createdAt: media.createdAt.toISOString(),
+        type: media.type,
+        url: media.url,
+        mimeType: media.mimeType,
+        width: media.width,
+        height: media.height,
+        duration: media.duration,
+        blurHash: media.blurHash,
+      })),
+    },
+  };
+}
