+++ apps/cms/src/types/dashboard.ts
@@ -1,4 +1,53 @@
+import type { ApiScope } from "@/utils/keys";
 import type { Media, MediaType } from "./media";
+
+export interface Category {
+  id: string;
+  name: string;
+  slug: string;
+  description?: string | null;
+  postsCount: number;
+}
+
+export interface Tag {
+  id: string;
+  name: string;
+  slug: string;
+  description?: string | null;
+  postsCount: number;
+}
+
+export interface Post {
+  id: string;
+  title: string;
+  coverImage: string | null;
+  status: "published" | "draft";
+  featured: boolean;
+  publishedAt: Date;
+  updatedAt: Date;
+  category: {
+    id: string;
+    name: string;
+  };
+  authors: Array<{
+    id: string;
+    name: string;
+    image: string | null;
+  }>;
+}
+
+export interface APIKey {
+  id: string;
+  name: string;
+  preview: string;
+  type: "public" | "private";
+  scopes: ApiScope[];
+  requestCount: number;
+  enabled: boolean;
+  lastUsed: Date | null;
+  expiresAt: Date | null;
+  createdAt: Date;
+}
 
 type DashboardRecentUpload = Pick<
   Media,
